BIGTOP-100. Whirr packaging build fails if we're not running the tests.
This looks to be a bit of weird legacy nonsense - cli/whirr.log is
only generated if we run the unit tests, but the only reason we care
about it is so that we can exclude it from the package in the first
place. For now, the easiest thing to do is to run the tests again.

git-svn-id: https://svn.apache.org/repos/asf/incubator/bigtop/trunk@1173183 13f79535-47bb-0310-9956-ffa450edef68